fancywm + remove kmonad

+++ .config/powershell/profile.ps1
@@ -10,12 +10,14 @@ Set-Alias open ii
 Invoke-Expression (& starship init powershell --print-full-init | Out-String)
 Invoke-Expression (& { (zoxide init powershell | Out-String) })
 
-function kbd {
-  ~/.config/kmonad/kmonad.exe "$([Environment]::GetEnvironmentVariable('userprofile'))/.config/kmonad/kmonad.kbd"
-}
+New-Variable -Name "ConfigPath" -Visibility Public -Value "$([Environment]::GetEnvironmentVariable('userprofile'))\.config"
 
-function whkd_reload {
-  taskkill /f /im whkd.exe
+
+function whkd_restart {
+  if (Get-Process whkd -ErrorAction SilentlyContinue)
+  {
+    taskkill /f /im whkd.exe
+  }
   Start-Process whkd -WindowStyle hidden
 }
 
@@ -24,20 +26,24 @@ function setup {
     $PowerShell,
     $WindowsTerminalPath,
     $PowerToys,
-    $RawAccel
+    $RawAccel,
+    $FancyWMPath
   )
   if ($PowerShell) {
-    sudo ln -s "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\powershell\profile.ps1" "$([Environment]::GetEnvironmentVariable('userprofile'))\Documents\WindowsPowerShell\Microsoft.PowerShell_profile.ps1"
+    sudo ln -s "$ConfigPath\powershell\profile.ps1" "$([Environment]::GetEnvironmentVariable('userprofile'))\Documents\WindowsPowerShell\Microsoft.PowerShell_profile.ps1"
   }
   if ($WindowsTerminalPath) {
     Remove-Item "$WindowsTerminalPath\settings.json"
-    sudo ln -s "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\windows-terminal\settings.json" "$WindowsTerminalPath\settings.json"
+    sudo ln -s "$ConfigPath\windows-terminal\settings.json" "$WindowsTerminalPath\settings.json"
   }
   if ($PowerToys) {
-    sudo ln -s "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\powertoys\keyboard-manager\settings.json" "$([Environment]::GetEnvironmentVariable('userprofile'))\AppData\Local\Microsoft\PowerToys\Keyboard Manager\settings.json"
-    sudo ln -s "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\powertoys\keyboard-manager\default.json" "$([Environment]::GetEnvironmentVariable('userprofile'))\AppData\Local\Microsoft\PowerToys\Keyboard Manager\default.json"
+    sudo ln -s "$ConfigPath\powertoys\keyboard-manager\settings.json" "$([Environment]::GetEnvironmentVariable('userprofile'))\AppData\Local\Microsoft\PowerToys\Keyboard Manager\settings.json"
+    sudo ln -s "$ConfigPath\powertoys\keyboard-manager\default.json" "$([Environment]::GetEnvironmentVariable('userprofile'))\AppData\Local\Microsoft\PowerToys\Keyboard Manager\default.json"
   }
   if ($RawAccel) {
-    sudo ln -s "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\rawaccel\settings.json" "$([Environment]::GetEnvironmentVariable('userprofile'))\.config\rawaccel\RawAccel\settings.json"
+    sudo ln -s "$ConfigPath\rawaccel\settings.json" "$ConfigPath\rawaccel\RawAccel\settings.json"
+  }
+  if ($FancyWMPath) {
+    sudo ln -s"$ConfigPath\fancywm\settings.json" "$FancyWMPath\settings.json"
   }
 }
